some issue fixed

+++ src/Page/MyListPage/MyListPage.jsx
@@ -1,4 +1,5 @@
 import { useContext, useEffect, useState } from "react";
+import Swal from "sweetalert2";
 import { AuthContext } from "../../Provider/AuthProvider";
 import MyListTable from "../../component/MyListTable";
 
@@ -7,23 +8,43 @@ const MyListPage = () => {
   const [myPlaces, setMyPlaces] = useState();
 
   useEffect(() => {
-    fetch(`http://localhost:7000/myList/${user?.email}`)
+    fetch(
+      `https://southeast-asia-server-kappa.vercel.app/myList/${user?.email}`
+    )
       .then((res) => res.json())
       .then((data) => {
         setMyPlaces(data);
       });
   }, [user?.email]);
   const handleDelete = (id) => {
-    console.log(id);
-    fetch(`http://localhost:7000/myPlaces/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data);
-        const remaining = myPlaces.filter((p) => p._id !== id);
-        setMyPlaces(remaining);
-      });
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`https://southeast-asia-server-kappa.vercel.app/myPlaces/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount > 0) {
+              console.log(data);
+              const remaining = myPlaces.filter((p) => p._id !== id);
+              setMyPlaces(remaining);
+              Swal.fire({
+                title: "Deleted!",
+                text: "Your file has been deleted.",
+                icon: "success",
+              });
+            }
+          });
+      }
+    });
   };
   return (
     <>
